[12/14 이재현] 블록, 트렉젝션 저장 기능 및 소스 정

+++ src/main/java/com/brchain/core/service/DockerService.java
@@ -8,34 +8,33 @@ import org.json.simple.JSONArray;
 import org.json.simple.JSONObject;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.brchain.core.client.DockerClient;
 import com.brchain.core.client.SshClient;
+import com.brchain.core.dto.ConInfoDto;
 import com.brchain.core.dto.ResultDto;
 import com.brchain.core.entity.ConInfoEntity;
 import com.brchain.core.util.Util;
 import com.google.common.collect.ImmutableList;
+import com.spotify.docker.client.exceptions.DockerException;
 import com.spotify.docker.client.messages.Container;
 import com.spotify.docker.client.messages.Container.PortMapping;
 
+import lombok.RequiredArgsConstructor;
+
 @Service
+@RequiredArgsConstructor
 public class DockerService {
 
+	private final DockerClient dockerClient;
+	private final SshClient sshClient;
+
+	private final ContainerService containerService;
+
+	private final Util util;
+
 	private Logger logger = LoggerFactory.getLogger(this.getClass());
-
-	@Autowired
-	private DockerClient dockerClient;
-
-	@Autowired
-	private ContainerService containerService;
-
-	@Autowired
-	private SshClient sshClient;
-
-	@Autowired
-	private Util util;
 
 	/**
 	 * 모든 컨테이너 삭제 서비스
@@ -54,22 +53,26 @@ public class DockerService {
 				Container container = iter.next();
 
 				ConInfoEntity conInfoEntity = null;
-				String path = "";
 
 				try {
-					conInfoEntity = containerService.removeConInfo(container.id());
+					conInfoEntity = containerService.deleteConInfo(container.id());
 				} catch (Exception e) {
 					logger.info("디비에 없는 컨테이너");
 				}
 
 				logger.info("[컨테이너 삭제] 컨테이너 id : " + container.id());
 				dockerClient.removeContainer(container.id());
-				sshClient.removeDir(conInfoEntity.getOrgName(), conInfoEntity.getConName());
+
+				if (conInfoEntity != null) {
+					sshClient.removeDir(conInfoEntity.getOrgName(), conInfoEntity.getConName());
+				}
 
 			}
 
 		} catch (Exception e) {
 
+			logger.error(e.getMessage());
+			e.printStackTrace();
 			return util.setResult("9999", false, e.getMessage(), null);
 
 		}
@@ -94,7 +97,7 @@ public class DockerService {
 
 			try {
 
-				conInfoEntity = containerService.removeConInfo(conId);
+				conInfoEntity = containerService.deleteConInfo(conId);
 
 			} catch (Exception e) {
 
@@ -108,6 +111,8 @@ public class DockerService {
 
 		} catch (Exception e) {
 
+			logger.error(e.getMessage());
+			e.printStackTrace();
 			return util.setResult("9999", false, e.getMessage(), null);
 
 		}
@@ -133,10 +138,10 @@ public class DockerService {
 				Container container = iter.next();
 
 				ConInfoEntity conInfoEntity = null;
-				String path = "";
+
 				if (container.names().get(0).contains(orgName)) {
 					try {
-						conInfoEntity = containerService.removeConInfo(container.id());
+						conInfoEntity = containerService.deleteConInfo(container.id());
 					} catch (Exception e) {
 						logger.info("디비에 없는 컨테이너");
 					}
@@ -150,6 +155,8 @@ public class DockerService {
 
 		} catch (Exception e) {
 
+			logger.error(e.getMessage());
+			e.printStackTrace();
 			return util.setResult("9999", false, e.getMessage(), null);
 
 		}
@@ -173,8 +180,6 @@ public class DockerService {
 			List<Container> containerList = dockerClient.loadAllContainers();
 
 			for (Container container : containerList) {
-
-				container.id();
 
 				JSONObject resultJson = new JSONObject();
 				ImmutableList<PortMapping> portList = container.ports();
@@ -200,11 +205,29 @@ public class DockerService {
 
 		} catch (Exception e) {
 
-			return util.setResult("9999", false, "e.getMessage()", null);
+			logger.error(e.getMessage());
+			e.printStackTrace();
+			return util.setResult("9999", false, e.getMessage(), null);
 
 		}
 
-		return util.setResult("0000", true, "Success get container info", resultJsonArr);
+		return util.setResult("0000", true, "Success get all containers info", resultJsonArr);
 	}
 
+	/**
+	 * conDto 샘성 함수
+	 * 
+	 * @param createConDto
+	 * 
+	 * @return 생성한 conDto
+	 * 
+	 * @throws DockerException
+	 * @throws InterruptedException
+	 */
+
+	public ConInfoDto createContainer(ConInfoDto createConDto) throws DockerException, InterruptedException {
+
+		return dockerClient.createContainer(createConDto);
+
+	}
 }
